bin/python-scrawler/requirements.txt
# MongoDB URL Scraper Dependencies
playwright>=1.40.0
pymongo>=4.0.0
python-dotenv>=1.0.0

# Note: The following are Python standard library modules (no installation needed):
# - asyncio
# - json  
# - datetime
# - urllib.parse
# - time
# - sys
# - os
